+++ src-docs/src/views/drag_and_drop/drag_and_drop_example.js
@@ -199,15 +199,18 @@ export const DragAndDropExample = {
         <React.Fragment>
           <p>
             By default the entire element surface can initiate a drag. To
-            specify a certain element within as the handle, set
-            <EuiCode>customDragHandle=true</EuiCode> on the{' '}
+            specify an element within as the handle and create a containing
+            group, set
+            <EuiCode>customDragHandle=true</EuiCode> and{' '}
+            <EuiCode>draggableContainer=true</EuiCode> on the{' '}
             <strong>EuiDraggable</strong>.
           </p>
           <p>
             The <EuiCode>provided</EuiCode> parameter on the{' '}
             <strong>EuiDraggable</strong> <EuiCode>children</EuiCode> render
             prop has all data required for functionality. Along with the{' '}
-            <EuiCode>customDragHandle</EuiCode> flag,
+            <EuiCode>customDragHandle</EuiCode> and{' '}
+            <EuiCode>draggableContainer</EuiCode>flags,
             <EuiCode>provided.dragHandleProps</EuiCode> needs to be added to the
             intended handle element.
           </p>
@@ -227,10 +230,12 @@ export const DragAndDropExample = {
         <React.Fragment>
           <p>
             <strong>EuiDraggable</strong> elements can contain interactive
-            elements such as buttons and form fields by adding the
-            <EuiCode>disableInteractiveElementBlocking</EuiCode> prop. This will
-            keep drag functionality while also enabling click, etc., events on
-            the interactive child elements.
+            elements such as buttons and form fields. Interactive elements
+            require <EuiCode>customDragHandle=true</EuiCode> and{' '}
+            <EuiCode>draggableContainer=true</EuiCode> on the{' '}
+            <strong>EuiDraggable</strong>. These props will maintain drag
+            functionality and accessibility, while enabling click, keypress,
+            etc., events on the interactive child elements.
           </p>
         </React.Fragment>
       ),